pr. an.: le script gère mieux l’ùnicode

+++ projets/analogique_2A/getipynb.sh
@@ -6,5 +6,11 @@ then
     exit 1
 fi
 
-~/scripts/ipynbtotex.py ~/Downloads/notebook.ipynb > tex/$1-genere.tex
+if [[ ! -f ~/Downloads/notebook.ipynb ]]
+then
+    echo 'Pas de notebook.ipynb'
+    exit 2
+fi
+
+~/scripts/ipynbtotex.py ~/Downloads/notebook.ipynb $1
 rm ~/Downloads/notebook.ipynb
